examples/inferencing/AutoMLOps/components/component_base/requirements.txt
explainable_ai_sdk
google-cloud-aiplatform
google-cloud-bigquery
kfp<2.0.0
